pyproject.toml
[project]
name = "photo-share"
version = "0.1.0"
description = ""
authors = [
    {name = "karkusha8",email = "chystykoleksandr@gmail.com"}
]
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-multipart (>=0.0.26,<0.0.27)",
    "bcrypt (==3.2.2)",
    "qrcode (>=8.2,<9.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "cloudinary (>=1.44.2,<2.0.0)",
    "pytest (>=9.0.3,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
